Apparently the database in here now is for the 3.3.5 clie
nt version server under the development branch in the server repo.  Added the 3.0.9 file from the serverOne repo, although I discovered this archive has an empty creatures table, so I added one I found where the structure of the table looked the same, but not sure of the data how 309 it actually is.